+++ libs/core/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.1 (2026-04-24)
+
+### 🩹 Fixes
+
+- **libs:** force rollup output.interop=auto on CJS bundle ([3480fdc](https://github.com/louiskhenghao/formzk/commit/3480fdc))
+
+### ❤️ Thank You
+
+- louiskhenghao @louiskhenghao
+
 ## 1.1.0 (2026-04-25)
 
 ### 🚀 Features